render markdown and plainify descriptions on previews
